feat(mf): respect APF v14 for discovering secondaries
before, the folders of the shared packages have been
searched for subfolders with a package.json. These
subfolders were added as secondary entry points.

Now, if the package.json contains an exports field,
like defined by APF v14, this field is used to find
out about secondary entry points. If this is not
the case, the original logic described aboved is
used as a fallback.