feat(nf): add build flag with options default and separate
before, all deps were build together to improve build performance.
now, you can set build to separate for selected shared deps. In this
case they will be built separate. This is necessary in rare cases, e. g. where a lib imports parts of it self using it's own package name. As native
federation uses externals such imports would not be bundled in when
built together with other deps. However, if we build it separately,
we can remove itself from the external list for this one build.